backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.4
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework==3.15.1


